+++ previous-versions/banking/pkg/models/operations/getaccount.go
@@ -42,6 +42,8 @@ type GetAccountResponse struct {
 	Account *shared.Account
 	// HTTP response content type for this operation
 	ContentType string
+	// Your API request was not properly authorized.
+	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -60,6 +62,13 @@ func (o *GetAccountResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
+}
+
+func (o *GetAccountResponse) GetErrorMessage() *shared.ErrorMessage {
+	if o == nil {
+		return nil
+	}
+	return o.ErrorMessage
 }
 
 func (o *GetAccountResponse) GetStatusCode() int {